IT WORKED 🥂 🎉 need to fix the lens
Hopefully this is it

I learned a bunch with this and the deployment was nowhere near as simple as I anticipated.

Getting to grips with react was not too bad, however coming back to javascript was really unintuitive for me after so long not using it. I used ChatGPT to translate my python function from the Docker API which essentially had the logic for the word counting down. However, there were a bunch more formatting and regexes that I needed to be able to treat dates, hyphenated words and apostrophe containing words as one word.

I tested everything locally and the functionality was good. I thought I would just be able to deploy to GitHub pages as I've done in the past, but since I wanted a realtime nature of the text input instead of a submit button, I needed to run an express node server to handle the POST requests. This meant I couldn't use a static page and it was my first time not just with React, but trying to deploy a hosted page with a server.

I struggled a lot with this and couldn't understand why my API was giving 405 despite everything seemingly checking out.

The trick was having this be a locally relative path for the project instead of using the hosted url. I still use the vercel url in the fetch POST response of handling text entry to the React TextInput widget, but the server and TextInput async response needed to be relative paths (/api/count-words)

I also had a bunch of problems with the vercel.json but this ended up being the key for the pathing needed for the serverless deployment.

Using POSTMAN helped me figure out that my response submissions and returns were not working correctly both locally and hosted via vercel. Everything was good when using the localhost via the browser but the POST request via POSTMAN was giving the same 405.

What I learned:
- I hate vercel.
- I need to learn more about non-static deployment.
- I need to completely revisit Javascript instead of yoloing it.
- I need to be able to comfortably write logic in JS
- Node is pretty strong when you get to grips with it and is now something I would consider for an option when I can see the benefits of completely carrying a project using it.